Change ContextClassLoader to NarClassLoader in AdditionalServlet
